Fix: headers should consume newline (#11)
Headers should consume all newlines preceding it (however it's not required). Double checked with VS Code's markdown preview to get same behaviour.

Example
```
Title


# header
```
should generate
![image](https://user-images.githubusercontent.com/5618601/42347955-685ce0b6-805c-11e8-8661-9b934d77a1ae.png)

### Bugfix
Found an issue where, if we were using `HeaderLineClassedRule` it wasn't rendering if the header didn't have multiple children. Updated the tests to use `HeaderLineClassedRule` by default. We already had a standard test for this case, but it wasn't caught because we were using the default `HeaderLineRule` and not the classed version. However, given the classed version being a super set I think it's okay to use that for testing purposes. We also use the classed rule in our actual app anyways.