Add state passing mechanism for parse specs
This allows setting of arbitrary state for child rules. In most cases rules can just pass on the state unchanged, but some rules, such as block quotes, will need the ability to not re-parse quotes inside of a quote, which can be achieved with this state. In these cases, the rule's `parse` method can create a modified copy of the state object to put on the returned `ParseSpec`. We therefore only end up creating new objects when the state actually changes.